Split out Status into a separate file.
