+++ process/info.go
@@ -7,32 +7,6 @@ import (
 	"gopkg.in/juju/charm.v6-unstable"
 )
 
-// Status values specific to workload processes.
-const (
-	StatusPending Status = iota
-	StatusActive
-	StatusFailed
-	StatusStopped
-)
-
-// Status represents the status of a worload process.
-type Status string
-
-// String implements fmt.Stringer.
-func (s Status) String() string {
-	switch status {
-	case StatusPending:
-		return "pending"
-	case StatusActive:
-		return "active"
-	case StatusFailed:
-		return "failed"
-	case StatusStopped:
-		return "stopped"
-	}
-	return "Unknown"
-}
-
 // ProcessInfo holds information about a process that Juju needs.
 type ProcessInfo struct {
 	charm.Process
